refactor(line): section에 대한 일급 컬렉션 객체인 Sections 생성 및 책임 분리
